add deprecated suffix
