+++ test/deprecated/CMakeLists.txt
@@ -150,6 +150,7 @@ if(WITH_TESTING)
 
   if(WITH_DISTRIBUTE)
     add_subdirectory(collective)
+    add_subdirectory(distributed_passes)
   endif()
 
 endif()
