+++ test/deprecated/distributed_passes/auto_parallel_pass_test_base_deprecated.py
@@ -18,13 +18,14 @@ import sys
 from collections import OrderedDict
 
 import numpy as np
-from dist_pass_test_base import DistPassTestBase
+from dist_pass_test_base_deprecated import DistPassTestBase
 
 import paddle
 from paddle.distributed import fleet
 from paddle.distributed.fleet import auto
 
-sys.path.append("../legacy_test")
+sys.path.append("../../legacy_test")
+
 import auto_parallel_gpt_model as modeling
 from auto_parallel_gpt_model import (
     GPTForPretraining,
