+++ test/deprecated/distributed_passes/pass_run_main_deprecated.py
@@ -17,7 +17,7 @@ import importlib
 import os
 import pickle
 
-from dist_pass_test_base import (
+from dist_pass_test_base_deprecated import (
     DistPassTestBase,
     prepare_python_path_and_return_module,
 )
