+++ test/deprecated/distributed_passes/test_auto_parallel_amp_pass_deprecated.py
@@ -16,17 +16,17 @@ import random
 import unittest
 
 import numpy as np
-from auto_parallel_pass_test_base import AutoParallelPassTestBase
+from auto_parallel_pass_test_base_deprecated import AutoParallelPassTestBase
 
 import paddle
 from paddle.distributed import fleet
 
 
-class TestRecomputePass(AutoParallelPassTestBase):
+class TestAMPPass(AutoParallelPassTestBase):
     def init(self):
         if paddle.is_compiled_with_cuda():
             paddle.set_flags({'FLAGS_cudnn_deterministic': 1})
-        self.rtol = 1e-6
+        self.rtol = 1e-5
         self.atol = 1e-8
 
         rank = paddle.distributed.get_rank()
@@ -36,17 +36,16 @@ class TestRecomputePass(AutoParallelPassTestBase):
 
     def apply_passes(self):
         dist_strategy = fleet.DistributedStrategy()
-        dist_strategy.recompute = True
-        dist_strategy.recompute_configs = {
-            "checkpoints": ["tmp_3", "tmp_6"],
-            "refined_ops_patterns": [
-                {
-                    "main_ops": ["matmul_v2", "elementwise_add"],
-                    "num": -1,
-                    "pre_ops": [],
-                    "suf_ops": [],
-                }
+        dist_strategy.amp = True
+        dist_strategy.amp_configs = {
+            "custom_white_list": [
+                'softmax',
+                'layer_norm',
+                'gelu',
             ],
+            "custom_black_list": ['c_softmax_with_cross_entropy'],
+            "init_loss_scaling": 32768,
+            "use_dynamic_loss_scaling": True,
         }
         dist_strategy.semi_auto = True
         fleet.init(is_collective=True, strategy=dist_strategy)
@@ -59,13 +58,6 @@ class TestRecomputePass(AutoParallelPassTestBase):
     def get_model(self, place, batch_size, sequence_len, vocab_size):
         return self.get_gpt_model(
             "mp", place, batch_size, sequence_len, vocab_size
-        )
-
-
-class TestRecomputePassDP(TestRecomputePass):
-    def get_model(self, place, batch_size, sequence_len, vocab_size):
-        return self.get_gpt_model(
-            "dp", place, batch_size, sequence_len, vocab_size
         )
