+++ test/deprecated/distributed_passes/test_auto_parallel_fp16_pass_deprecated.py
@@ -16,7 +16,7 @@ import random
 import unittest
 
 import numpy as np
-from auto_parallel_pass_test_base import AutoParallelPassTestBase
+from auto_parallel_pass_test_base_deprecated import AutoParallelPassTestBase
 
 import paddle
 from paddle.distributed import fleet
