+++ test/deprecated/distributed_passes/test_dist_fuse_sgd_pass_deprecated.py
@@ -16,7 +16,7 @@ import os
 import unittest
 
 import numpy as np
-from dist_pass_test_base import DistPassTestBase
+from dist_pass_test_base_deprecated import DistPassTestBase
 
 import paddle
 from paddle import nn
@@ -30,24 +30,22 @@ class DemoNet(nn.Layer):
     def __init__(self):
         super().__init__()
 
-        self.conv1 = nn.Conv2D(3, 3, (3, 3), padding=1, data_format="NHWC")
-        self.conv2 = nn.Conv2D(3, 3, (3, 3), padding=1, data_format="NHWC")
-        self.conv3 = nn.Conv2D(3, 3, (3, 3), padding=1, data_format="NHWC")
+        self.conv1 = nn.Conv2D(3, 8, (3, 3), data_format="NHWC")
+        self.bn1 = nn.BatchNorm2D(8, data_format="NHWC")
         self.relu = nn.ReLU()
 
     def forward(self, x):
         out = self.conv1(x)
-        out = self.conv2(out) + self.conv3(out)
+        out = self.bn1(out)
         out = self.relu(out)
         out = paddle.flatten(out, 1)
         return out
 
 
-class TestInplaceAddtoPass(DistPassTestBase):
+class TestFuseAdamPass(DistPassTestBase):
     def init(self):
-        self.atol = 0.0
-        self.rtol = 0.0
-        paddle.base.set_flags({"FLAGS_max_inplace_grad_add": 8})
+        self.atol = 1e-4
+        self.rtol = 1e-4
 
     def get_model(self, place, batch_size=32, image_shape=[224, 224, 3]):
         image = paddle.static.data(
@@ -57,12 +55,11 @@ class TestInplaceAddtoPass(DistPassTestBase):
         model = DemoNet()
         pred_out = model(image)
         loss = paddle.mean(pred_out)
-        optimizer = paddle.optimizer.Adam(learning_rate=1e-3)
+        optimizer = paddle.optimizer.SGD(learning_rate=1e-3)
 
         dist_strategy = fleet.DistributedStrategy()
         dist_strategy.fuse_all_reduce_ops = False
         dist_strategy.without_graph_optimization = True
-
         fleet.init(is_collective=True, strategy=dist_strategy)
         optimizer = fleet.distributed_optimizer(optimizer)
         optimizer.minimize(loss)
@@ -78,23 +75,28 @@ class TestInplaceAddtoPass(DistPassTestBase):
 
         main_program = paddle.static.default_main_program()
         startup_program = paddle.static.default_startup_program()
-        print(main_program)
         return main_program, startup_program, [image], [loss], reader
 
     def apply_passes(self, main_prog, startup_prog):
-        pass_manager = PassManager(
-            [new_pass("inplace_addto_op", {"use_cuda": True})]
-        )
+        pass_manager = PassManager([new_pass("fuse_optimizer")])
         pass_manager.apply([main_prog], [startup_prog])
         print(pass_manager.names)
 
-        conv2d_grad_attr = []
+        op_type = []
         for op in main_prog.global_block().ops:
-            if op.type == "conv2d_grad":
-                conv2d_grad_attr.append(op.desc.attr("use_addto"))
-        self.assertTrue(True in conv2d_grad_attr)
+            op_type.append(op.type)
+            if op.type == "sgd":
+                self.assertTrue(
+                    "@FUSEDVAR@_sgd_Param_batch_norm2d_0.b_0"
+                    in op.input("Param")
+                )
+                self.assertTrue(
+                    "@FUSEDVAR@_sgd_Grad_batch_norm2d_0.b_0@GRAD"
+                    in op.input("Grad")
+                )
+        self.assertTrue("coalesce_tensor" in op_type)
 
-    def test_inplace_addto(self):
+    def test_fuse_adam(self):
         self.check_main()
