+++ test/distributed_passes/check_pass_conflict_example.py
@@ -12,9 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import sys
 import unittest
 
-from dist_pass_test_base import PassConflictChecker
+sys.path.append("../deprecated/distributed_passes")
+
+from dist_pass_test_base_deprecated import PassConflictChecker
 from model_zoo import resnet_model
 
 from paddle.distributed.passes import new_pass
