+++ test/distributed_passes/test_build_cinn_pass_resnet.py
@@ -12,9 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import sys
 import unittest
 
-from dist_pass_test_base import DistPassTestBase
+sys.path.append("../deprecated/distributed_passes")
+
+from dist_pass_test_base_deprecated import DistPassTestBase
 
 from paddle.distributed.passes import PassManager, new_pass
