+++ test/distributed_passes/test_dist_fuse_all_reduce_pass.py
@@ -12,9 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import sys
 import unittest
 
-from dist_pass_test_base import DistPassTestBase
+sys.path.append("../deprecated/distributed_passes")
+from dist_pass_test_base_deprecated import DistPassTestBase
 from model_zoo import resnet_model
 
 from paddle.distributed.passes import PassManager, new_pass
