+++ test/distributed_passes/test_dist_fuse_gemm_epilogue_pass.py
@@ -12,10 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import sys
 import unittest
 
 import numpy as np
-from dist_pass_test_base import DistPassTestBase
+
+sys.path.append("../deprecated/distributed_passes")
+from dist_pass_test_base_deprecated import DistPassTestBase
 
 import paddle
 from paddle import nn
