+++ tools/test_run_by_protobuf_3.py
@@ -48,13 +48,13 @@ PROTOBUF3_NEEDED_TESTING_LIST = [
     'test_dist_sparse_tensor_load_adam',
     'test_dist_sparse_tensor_load_sgd_deprecated',
     'test_communicator_sync',
-    'test_dist_fuse_adam_pass',
-    'test_dist_fuse_bn_act_pass',
-    'test_dist_fuse_bn_add_act_pass',
-    'test_dist_fuse_momentum_pass',
-    'test_dist_fuse_relu_depthwise_conv_pass',
-    'test_dist_fuse_sgd_pass',
-    'test_dist_inplace_addto_pass',
+    'test_dist_fuse_adam_pass_deprecated',
+    'test_dist_fuse_bn_act_pass_deprecated',
+    'test_dist_fuse_bn_add_act_pass_deprecated',
+    'test_dist_fuse_momentum_pass_deprecated',
+    'test_dist_fuse_relu_depthwise_conv_pass_deprecated',
+    'test_dist_fuse_sgd_pass_deprecated',
+    'test_dist_inplace_addto_pass_deprecated',
     'test_fleet_sharding_meta_optimizer',
     'test_rnn_dp',
     'test_communicator_half_async',
